+++ Tests/KeyboardKitTests/Actions/KeyboardAction+AliasesTests.swift
@@ -1,5 +1,5 @@
 //
-//  KeyboardActionRowTest.swift
+//  KeyboardAction+AliasesTests.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2019-07-04.
@@ -10,7 +10,7 @@ import Quick
 import Nimble
 import KeyboardKit
 
-class KeyboardActionRowTest: QuickSpec {
+class KeyboardAction_AliasesTests: QuickSpec {
     
     override func spec() {